+++ source/isaaclab_physx/isaaclab_physx/sensors/pva/pva_data.py
@@ -78,7 +78,9 @@ class PvaData(BasePvaData):
 
     @property
     def lin_acc_b(self) -> wp.array:
-        """PVA frame linear acceleration relative to the world expressed in PVA frame.
+        """Linear acceleration (coordinate) in the PVA frame [m/s^2].
+
+        Equal to -g in freefall, zero at rest.
 
         Shape is (num_instances,), dtype = wp.vec3f. In torch this resolves to (num_instances, 3).
         """